fix: close storefront parity gaps
