+++ packages/cli/package.json
@@ -41,6 +41,7 @@
   "license": "MIT",
   "dependencies": {},
   "devDependencies": {
+    "@lucid-agents/types": "workspace:*",
     "@types/node": "^24.3.1",
     "bun-types": "^1.1.16",
     "tsup": "^8.5.0",
