+++ packages/cli/templates/trading-data-agent/agent.ts.template
@@ -1,18 +1,21 @@
 {{TEMPLATE_IMPORTS}}
-import { paymentsFromEnv } from '@lucid-agents/payments';
+import { ap2 } from '@lucid-agents/ap2';
+import { createAgent } from '@lucid-agents/core';
+import { http } from '@lucid-agents/http';
+import { payments, paymentsFromEnv } from '@lucid-agents/payments';
 
-const payments = paymentsFromEnv();
+const paymentConfig = paymentsFromEnv();
 
 {{TEMPLATE_PRE_SETUP}}
-const appOptions = {
-  payments: payments || {
-    facilitatorUrl: process.env.PAYMENTS_FACILITATOR_URL as `${string}://${string}`,
-    facilitatorAuth: process.env.PAYMENTS_FACILITATOR_AUTH,
-    payTo: process.env.PAYMENTS_RECEIVABLE_ADDRESS as `0x${string}`,
-    network: process.env.PAYMENTS_NETWORK as any,
-  },
-  ap2: payments ? { roles: ['merchant'], required: true } : undefined,
-};
+const agent = await createAgent({
+  name: process.env.AGENT_NAME ?? 'trading-data-agent',
+  version: process.env.AGENT_VERSION ?? '0.1.0',
+  description: process.env.AGENT_DESCRIPTION ?? 'Trading data provider agent',
+})
+  .use(http())
+  .use(payments({ config: paymentConfig }))
+  .use(ap2({ roles: ['merchant'], required: true }))
+  .build();
 
 {{TEMPLATE_POST_SETUP}}
 
@@ -34,14 +37,16 @@ addEntrypoint({
     price: z.number(),
     volume: z.number(),
     timestamp: z.string(),
-    data: z.array(z.object({
-      time: z.string(),
-      open: z.number(),
-      high: z.number(),
-      low: z.number(),
-      close: z.number(),
-      volume: z.number(),
-    })),
+    data: z.array(
+      z.object({
+        time: z.string(),
+        open: z.number(),
+        high: z.number(),
+        low: z.number(),
+        close: z.number(),
+        volume: z.number(),
+      })
+    ),
   }),
   price: '0.005', // USD decimal price for historical data
   handler: async ctx => {
