+++ packages/http/src/service-ui/define-service-ui.ts
@@ -1,0 +1,6 @@
+import type { ServiceUiConfig } from '@lucid-agents/types/http';
+
+/** Provides contextual typing for a generated service-ui.config.ts file. */
+export function defineServiceUi(config: ServiceUiConfig): ServiceUiConfig {
+  return config;
+}
